third_party/WebKit/LayoutTests/http/tests/security/window-named-proto-expected.txt

CONSOLE ERROR: line 1: Uncaught TypeError: Cannot read property 'innerHTML' of null
